diet_backend/resources/auth.py
import falcon
import jwt
import datetime
from falcon import HTTPUnauthorized

SECRET_KEY = 'your_secret_key'  # Use a secure key in production

class Auth:
    def register(self, req, resp):
        # Register logic here (save user details in DB)
        pass

    def login(self, req, resp):
        # Validate user and create a JWT token
        username = req.media.get('username')
        password = req.media.get('password')
        # Validate username and password
        if valid_user(username, password):
            token = jwt.encode({
                'user_id': username,
                'exp': datetime.datetime.now(datetime.timezone.utc) + datetime.timedelta(hours=1)
            }, SECRET_KEY, algorithm='HS256')
            resp.media = {'token': token}
        else:
            raise HTTPUnauthorized

    def logout(self, req, resp):
        # Invalidate token logic (if using server-side session)
        pass

# Validate user function (for example purposes)
def valid_user(username, password):
    # Implement user validation logic here
    return True
