diet_backend/requirements.txt
falcon==3.1.3
gunicorn==23.0.0
mysql-connector-python==8.0.29
tensorflow
numpy==1.24.1
scikit-learn==1.1.3
pandas==1.5.1
pydantic==1.10.4
pyjwt==1.7.1
sqlalchemy==2.0.34
